java oop, java inheritance, java methods and attributes

+++ src/com/mateusborja/javadevanalyst/lesson_4_4/Valor.java
@@ -2,7 +2,7 @@ package com.mateusborja.javadevanalyst.lesson_4_4;
 
 public class Valor {
 	
-	int v;
+	double v;
 	
 	int z;
